# server/src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Ballot.io",
    "description": "Real-time e-voting application API",
    "termsOfService": "https://www.google.com/",
    "contact": {
      "name": "Wolox Team",
      "email": "hello@wolox.co",
      "url": "https://www.wolox.com.ar/"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local server"
    },
    {
      "url": "https://api_url_testing",
      "description": "Testing server"
    },
    {
      "url": "https://api_url_production",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "tags": [
    {
      "name": "Auth operations"
    },
    {
      "name": "User operations"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth operations"],
        "description": "Register new users",
        "operationId": "registerUsers",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Email sent successfully to johndoe@gmail.com"
          },
          "401": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "message": "\"password\" is not allowed to be empty"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth operations"],
        "description": "Login a user",
        "operationId": "loginUsers",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully logged in",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eygwZjQwZGIiLCJV4cCI6MTYwNDQxMTI4NH0.4dv6IJ3NPwCxkk7WRwj8pehabO5lVxEEOVKLTTVLjvQ",
                  "user": {
                      "_id": "5fa07ae1335ecf11280f40db",
                      "name": "Kelechi Chinaka",
                      "email": "mrkelechichinaka@gmail.com"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "message": "\"password\" is not allowed to be empty"
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "put": {
        "tags": ["Auth operations"],
        "description": "Generate a new user password",
        "operationId": "passwordUsers",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "email": "johndoe@gmail.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password link sent to email",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Email has been sent to johndoe@gmail.com. Follow the instruction to set a new password."
                }
              }
            }
          },
          "401": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "\"email\" is required"
              }
              }
            }
          }
        }
      }
    },
    "/user/polls/userID/:userID": {
      "get": {
        "tags": ["User operations"],
        "description": "Get list of poll by userID",
        "operationId": "getUserPollList",
        "parameters": [
          {
            "name": "userID",
            "n": "header",
            "schema": {
              "$ref": "#/components/schemas/companyId"
            },
            "required": true,
            "description": "Add userID as query params, where /:userID is the _id of user"
          }
        ],
        "responses": {
          "200": {
            "description": "User list of poll provided",
            "content": {
              "application/json": {
                "example": {
                  "status": true,
                  "data": [{
                    "options": [{
                      "name": "APC",
                        "count": 0,
                        "option_id": "5fa16dc8cc405a084471a42c"
                      },
                      {
                        "name": "PDP",
                        "count": 0,
                        "option_id": "5fa16dc8cc405a084471a42d"
                      },
                      {
                        "name": "YouthFC",
                        "count": 0,
                        "option_id": "5fa16dc8cc405a084471a42e"
                      }],
                      "_id": "5fa16dc8cc405a084471a42f",
                      "createdBy": "5f92a956072dee19b07981b1",
                      "title": "Election2023",
                      "totalVotes": 0,
                      "date": "2020-11-03T14:48:40.840Z",
                      "__v": 0
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing required parameters. eg token",
            "content": {
              "application/json": {
                "example": {
                  "status": false,
                  "message": "No token provided. Access denied!"
                }
              }
            }
          }
        }
      }
    },
    "/user/poll/:pollID/delete": {
      "delete": {
        "tags": ["User operations"],
        "description": "Delete a poll.",
        "operationId": "deletePoll",
        "parameters": [
          {
            "name": "pollID",
            "n": "header",
            "schema": {
              "$ref": "#/components/schemas/companyId"
            },
            "required": true,
            "description": "Add pollID as query params, where /:pollID been the _id of the poll to be deleted."
          }
        ],
        "responses": {
          "200": {
            "description": "Delete successfull",
            "content": {
              "application/json": {
                "example": {
                  "status": true,
                  "message": "Successfully deleted item!"
                }
              }
            }
          }
        }
      }
    },
    "/user/poll/create": {
      "post": {
        "tags": ["User operations"],
        "description": "Create new poll",
        "operationId": "createPoll",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePollPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Poll created successfully",
            "content": {
              "application/json": {
                "example": {
                  "status": true,
                  "message": "Poll created successfully",
                  "_id": "5fa1a11490563a158cda4fd6"
                }
              }
            }
          },
          "401": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "example": {
                  "status": false,
                  "message": "\"createdBy\" is required"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "name": {
        "type": "string",
        "example": "John Doe"
      },
      "email": {
        "type": "string",
        "example": "johndoe@gmail.com"
      },
      "password": {
        "type": "string",
        "example": "123456"
      },
      "createdBy": {
        "type": "string",
        "example": "5fa19cef90563a158cda4fc2"
      },
      "title": {
        "type": "string",
        "example": "Election2023"
      },
      "options": {
        "type": "array",
        "example": ["APC", "PDP", "YouthFC"]
      },
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/name"
          },
          "email": {
            "$ref": "#/components/schemas/email"
          },
          "password": {
            "$ref": "#/components/schemas/password"
          }
        }
      },
      "Login": {
        "type": "object",
        "properties": {
          "email": {
            "$ref": "#/components/schemas/email"
          },
          "password": {
            "$ref": "#/components/schemas/password"
          }
        }
      },
      "CreatePollPayload": {
        "type": "object",
        "properties": {
          "createdBy": {
            "$ref": "#/components/schemas/createdBy"
          },
          "title": {
            "$ref": "#/components/schemas/title"
          },
          "options": {
            "$ref": "#/components/schemas/options"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": true
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}